Update the Amazon EC2 setup instructions
Primarily this was to add a line to rc.local that restarts the docker
image if the instance is stopped and started again. This was something
I presonally needed due to a constraint on the sandbox environment I
was setting up Juice Shop in, where instances were all stopped every
night.

I also added some comments to the script, and a note on how to set
the environment variables for CTF mode.